requirements.txt
streamlit
streamlit-authenticator
google-generativeai
python-dotenv
